Rename update tx
