+++ src/sync/index.js
@@ -8,7 +8,7 @@ const BigNumber = require('bignumber.js');
 const { getContractMetadata, isMainnet } = require('../config');
 const { BLOCK_0_TIMESTAMP, SATOSHI_CONVERSION } = require('../constants');
 const { db, DBHelper } = require('../db');
-const updateTxDB = require('./update_tx');
+const updateTxDB = require('./updateLocalTx');
 
 const Topic = require('../models/topic');
 const CentralizedOracle = require('../models/centralizedOracle');
